requirements.txt
asgiref==3.9.1
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
crispy-tailwind==1.0.3
Django==5.2.6
django-crispy-forms==2.4
django-environ==0.12.0
django-storages==1.14.6
djangorestframework==3.16.1
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-core==2.4.3
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
idna==3.10
proto-plus==1.26.1
protobuf==6.32.1
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
rsa==4.9.1
spotipy==2.25.1
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
